requirements.txt
# Cricket-API dependencies (pinned for compatibility with googlesearch-python)
Flask>=2.0,<4
requests>=2.28
beautifulsoup4>=4.9
googlesearch-python>=1.3.0
lxml>=4.9.0
selenium>=4.0